requirements.txt
streamlit
python-barcode
pillow
pandas
firebase-admin
